refactor: update ID fields to int64 across models and utility functions

+++ internal/domain/entity/cake.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Cake struct {
-	ID          int          `gorm:"column:id;primaryKey"`
+	ID          int64        `gorm:"column:id;primaryKey"`
 	Title       string       `gorm:"column:title"`
 	Description string       `gorm:"column:description"`
 	Price       float64      `gorm:"column:price"`
