+++ internal/domain/entity/customer.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Customer struct {
-	ID        int          `gorm:"column:id;primaryKey"`
+	ID        int64        `gorm:"column:id;primaryKey"`
 	Name      string       `gorm:"column:name"`
 	Email     string       `gorm:"column:email;unique"`
 	Password  string       `gorm:"column:password"`
