+++ internal/domain/model/wishlist.go
@@ -1,9 +1,9 @@
 package model
 
 type WishListResponse struct {
-	Cakes      []CakeModel `json:"cakes"`
+	Cakes []CakeModel `json:"cakes"`
 }
 
 type AddToWishListRequest struct {
-	CakeID int `json:"cake_id" validate:"required"`
+	CakeID int64 `json:"cake_id" validate:"required"`
 }
